Blue Side SkyStone Detection and Placement, Fixing turnToHeading(), and repositionFoundation()
-Needed to set the currentHeading() to getHeading() in the generalRoationControl()

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/GFORCE_Navigation.java
@@ -442,7 +442,7 @@ public class GFORCE_Navigation
     }
 
 
-    public boolean waitForTarget(int timeout) {
+    public boolean waitForTarget(double timeout) {
         navTime.reset();
         while (myOpMode.opModeIsActive() && !targetIsVisible(0) && (navTime.time() < timeout)) {
             showNavTelemetry(true);
